fix skip_empty_rows('skip') and trailing newlines
When reading line by line, Text::CSV returns an empty row if the input
file contains an empty line, even if skip_empty_rows('skip') is in use.

This happens due to the following code (greatly abridged here):

    $seenSomething = 0;
    LOOP:
        while($self->__get_from_src($ctx, $src)) {
            $seenSomething = 1;
            # (This conditional happens when EOL is reached and
            # skip_empty_row is set).
            if (!defined $c or $ser == 2) {  # EOF
                $waitingForField = 0;
                last LOOP;
            }
        }

    if ($waitingForField) {
        if ($seenSomething or !$ctx->{useIO}) {
            return 1;
        }
        $self->SetDiag(2012);
        return;
    }

The call to SetDiag is necessary to set the EOF condition (2012 is EOF).
Without this, __parse() returns a truthy value, so getline() returns an
empty array.

In order to make the SetDiag call happen, we need to:
* No longer unset $waitingForField (after skipping an empty row, we're
  still waiting).
* Unset $seenSomething (we did see something, but we skipped it).

This fixes:
https://github.com/makamaka/Text-CSV/issues/64

The corresponding change is now in Text::CSV_XS from:
https://github.com/Tux/Text-CSV_XS/pull/52

This change includes whitespace and test descriptions from:
https://github.com/Tux/Text-CSV_XS/commit/731e78c7b61ab245249ae50bf473ad8d742b0194